Move controller logic into internal package
This ensures that new providers won't risk having their logic imported as a
library unless they make the conscious decision to allow doing so.
